refactor: updated to new analysis, deleted old
