+++ analysis/requirements.in
@@ -1,9 +1,15 @@
+# Core dependencies
 duckdb>=0.9.2
+duckdb-engine>=0.9.2
+llama-index>=0.10.0
+sqlalchemy>=2.0.27
 openai>=1.12.0
-pydantic>=2.6.1
-rich>=13.7.0
-typer>=0.9.0
-pandas>=2.2.0
-pyarrow>=15.0.0
 boto3>=1.34.0
-aws-lambda-powertools>=2.32.0
+thefuzz[speedup]>=0.20.0
+
+# Data processing
+pandas>=2.2.0
+
+# UI dependencies
+streamlit>=1.31.1
+plotly>=5.19.0
