+++ analysis/requirements.txt
@@ -1,99 +1,335 @@
 # This file was autogenerated by uv via the following command:
 #    uv pip compile requirements.in -o requirements.txt
+aiohappyeyeballs==2.4.6
+    # via aiohttp
+aiohttp==3.11.12
+    # via llama-index-core
+aiosignal==1.3.2
+    # via aiohttp
+altair==5.5.0
+    # via streamlit
 annotated-types==0.7.0
     # via pydantic
-anyio==4.8.0
+anyio==3.7.1
     # via
     #   httpx
     #   openai
-aws-lambda-powertools==3.6.0
+attrs==25.1.0
+    # via
+    #   aiohttp
+    #   jsonschema
+    #   referencing
+beautifulsoup4==4.13.3
+    # via llama-index-readers-file
+blinker==1.9.0
+    # via streamlit
+boto3==1.36.26
     # via -r requirements.in
-boto3==1.36.24
-    # via -r requirements.in
-botocore==1.36.24
+botocore==1.36.26
     # via
     #   boto3
     #   s3transfer
+cachetools==5.5.2
+    # via streamlit
 certifi==2025.1.31
     # via
     #   httpcore
     #   httpx
+    #   llama-cloud
+    #   requests
+charset-normalizer==3.4.1
+    # via requests
 click==8.1.8
-    # via typer
+    # via
+    #   llama-cloud-services
+    #   nltk
+    #   streamlit
+dataclasses-json==0.6.7
+    # via llama-index-core
+deprecated==1.2.18
+    # via llama-index-core
+dirtyjson==1.0.8
+    # via llama-index-core
 distro==1.9.0
     # via openai
-duckdb==1.2.0
+duckdb==0.9.2
+    # via
+    #   -r requirements.in
+    #   duckdb-engine
+duckdb-engine==0.9.2
     # via -r requirements.in
+filetype==1.2.0
+    # via llama-index-core
+frozenlist==1.5.0
+    # via
+    #   aiohttp
+    #   aiosignal
+fsspec==2025.2.0
+    # via llama-index-core
+gitdb==4.0.12
+    # via gitpython
+gitpython==3.1.44
+    # via streamlit
+greenlet==3.1.1
+    # via sqlalchemy
 h11==0.14.0
     # via httpcore
 httpcore==1.0.7
     # via httpx
 httpx==0.28.1
-    # via openai
+    # via
+    #   llama-cloud
+    #   llama-index-core
+    #   openai
 idna==3.10
     # via
     #   anyio
     #   httpx
+    #   requests
+    #   yarl
+jinja2==3.1.5
+    # via
+    #   altair
+    #   pydeck
 jiter==0.8.2
     # via openai
 jmespath==1.0.1
     # via
-    #   aws-lambda-powertools
     #   boto3
     #   botocore
+joblib==1.4.2
+    # via nltk
+jsonschema==4.23.0
+    # via altair
+jsonschema-specifications==2024.10.1
+    # via jsonschema
+llama-cloud==0.1.13
+    # via
+    #   llama-cloud-services
+    #   llama-index-indices-managed-llama-cloud
+llama-cloud-services==0.6.1
+    # via llama-parse
+llama-index==0.12.19
+    # via -r requirements.in
+llama-index-agent-openai==0.4.6
+    # via
+    #   llama-index
+    #   llama-index-program-openai
+llama-index-cli==0.4.0
+    # via llama-index
+llama-index-core==0.12.19
+    # via
+    #   llama-cloud-services
+    #   llama-index
+    #   llama-index-agent-openai
+    #   llama-index-cli
+    #   llama-index-embeddings-openai
+    #   llama-index-indices-managed-llama-cloud
+    #   llama-index-llms-openai
+    #   llama-index-multi-modal-llms-openai
+    #   llama-index-program-openai
+    #   llama-index-question-gen-openai
+    #   llama-index-readers-file
+    #   llama-index-readers-llama-parse
+llama-index-embeddings-openai==0.3.1
+    # via
+    #   llama-index
+    #   llama-index-cli
+llama-index-indices-managed-llama-cloud==0.6.8
+    # via llama-index
+llama-index-llms-openai==0.3.20
+    # via
+    #   llama-index
+    #   llama-index-agent-openai
+    #   llama-index-cli
+    #   llama-index-multi-modal-llms-openai
+    #   llama-index-program-openai
+    #   llama-index-question-gen-openai
+llama-index-multi-modal-llms-openai==0.4.3
+    # via llama-index
+llama-index-program-openai==0.3.1
+    # via
+    #   llama-index
+    #   llama-index-question-gen-openai
+llama-index-question-gen-openai==0.3.0
+    # via llama-index
+llama-index-readers-file==0.4.5
+    # via llama-index
+llama-index-readers-llama-parse==0.4.0
+    # via llama-index
+llama-parse==0.6.1
+    # via llama-index-readers-llama-parse
 markdown-it-py==3.0.0
     # via rich
+markupsafe==3.0.2
+    # via jinja2
+marshmallow==3.26.1
+    # via dataclasses-json
 mdurl==0.1.2
     # via markdown-it-py
+multidict==6.1.0
+    # via
+    #   aiohttp
+    #   yarl
+mypy-extensions==1.0.0
+    # via typing-inspect
+narwhals==1.27.1
+    # via
+    #   altair
+    #   plotly
+nest-asyncio==1.6.0
+    # via llama-index-core
+networkx==3.4.2
+    # via llama-index-core
+nltk==3.9.1
+    # via
+    #   llama-index
+    #   llama-index-core
 numpy==2.2.3
-    # via pandas
+    # via
+    #   llama-index-core
+    #   pandas
+    #   pydeck
+    #   streamlit
 openai==1.63.2
-    # via -r requirements.in
-pandas==2.2.3
-    # via -r requirements.in
-pyarrow==19.0.1
-    # via -r requirements.in
-pydantic==2.10.6
     # via
     #   -r requirements.in
+    #   llama-index-agent-openai
+    #   llama-index-embeddings-openai
+    #   llama-index-llms-openai
+packaging==24.2
+    # via
+    #   altair
+    #   marshmallow
+    #   plotly
+    #   streamlit
+pandas==2.2.3
+    # via
+    #   -r requirements.in
+    #   llama-index-readers-file
+    #   streamlit
+pillow==11.1.0
+    # via
+    #   llama-index-core
+    #   streamlit
+plotly==6.0.0
+    # via -r requirements.in
+propcache==0.3.0
+    # via
+    #   aiohttp
+    #   yarl
+protobuf==5.29.3
+    # via streamlit
+pyarrow==19.0.1
+    # via streamlit
+pydantic==2.10.6
+    # via
+    #   llama-cloud
+    #   llama-cloud-services
+    #   llama-index-core
     #   openai
 pydantic-core==2.27.2
     # via pydantic
+pydeck==0.9.1
+    # via streamlit
 pygments==2.19.1
     # via rich
+pypdf==5.3.0
+    # via llama-index-readers-file
 python-dateutil==2.9.0.post0
     # via
     #   botocore
     #   pandas
+python-dotenv==1.0.1
+    # via llama-cloud-services
 pytz==2025.1
     # via pandas
-rich==13.9.4
+pyyaml==6.0.2
+    # via llama-index-core
+rapidfuzz==3.12.1
+    # via thefuzz
+referencing==0.36.2
     # via
-    #   -r requirements.in
-    #   typer
+    #   jsonschema
+    #   jsonschema-specifications
+regex==2024.11.6
+    # via
+    #   nltk
+    #   tiktoken
+requests==2.32.3
+    # via
+    #   llama-index-core
+    #   streamlit
+    #   tiktoken
+rich==13.9.4
+    # via streamlit
+rpds-py==0.23.1
+    # via
+    #   jsonschema
+    #   referencing
 s3transfer==0.11.2
     # via boto3
-shellingham==1.5.4
-    # via typer
 six==1.17.0
     # via python-dateutil
+smmap==5.0.2
+    # via gitdb
 sniffio==1.3.1
     # via
     #   anyio
     #   openai
-tqdm==4.67.1
-    # via openai
-typer==0.15.1
+soupsieve==2.6
+    # via beautifulsoup4
+sqlalchemy==2.0.27
+    # via
+    #   -r requirements.in
+    #   duckdb-engine
+    #   llama-index-core
+streamlit==1.42.2
     # via -r requirements.in
+striprtf==0.0.26
+    # via llama-index-readers-file
+tenacity==8.5.0
+    # via
+    #   llama-index-core
+    #   streamlit
+thefuzz==0.22.1
+    # via -r requirements.in
+tiktoken==0.9.0
+    # via llama-index-core
+toml==0.10.2
+    # via streamlit
+tornado==6.4.2
+    # via streamlit
+tqdm==4.67.1
+    # via
+    #   llama-index-core
+    #   nltk
+    #   openai
 typing-extensions==4.12.2
     # via
-    #   anyio
-    #   aws-lambda-powertools
+    #   altair
+    #   beautifulsoup4
+    #   llama-index-core
     #   openai
     #   pydantic
     #   pydantic-core
-    #   typer
+    #   referencing
+    #   sqlalchemy
+    #   streamlit
+    #   typing-inspect
+typing-inspect==0.9.0
+    # via
+    #   dataclasses-json
+    #   llama-index-core
 tzdata==2025.1
     # via pandas
 urllib3==2.3.0
-    # via botocore
+    # via
+    #   botocore
+    #   requests
+wrapt==1.17.2
+    # via
+    #   deprecated
+    #   llama-index-core
+yarl==1.18.3
+    # via aiohttp
