--- llama_query/requirements.in
@@ -1,7 +1,0 @@
-duckdb==0.9.2
-duckdb-engine==0.9.2
-llama-index>=0.10.0
-sqlalchemy==2.0.27
-openai>=1.12.0
-boto3>=1.34.0
-thefuzz[speedup]>=0.20.0
